add(traits): add the "readable" traits for implicit arrays
This development should add a set of traits specific to the implicit
backend/function behind an implicit array. The idea is to be able to
give an implicit array a wide variety of types of backend and to have it
work out of the box without respecting a strict class interface.