Special mapper for passing vtkmDataSet through for rendering
This change introduces a vtkmDataSetMapper whose role is to map
vtkmDataSets to graphics primitives. As of now, vtkmDataSetMapper is
just a no-op implementation that gets overriden when OSPRay is enabled.
In that case, the OSPRay view node factory replaces the mapper with the
vtkmOSPRayDataSetMapperNode that will hand over the vtkmDataSet directly
to OSPRay for rendering without offloading from the GPU device.