vtkCGNSReader: Read meshes with arbitrary polyhedra and canonical elements
This adds support to read CGNS meshes containing arbitrary polyhedra
(with the CGNS nodes NGon and NFace) and canonical elements (triangles,
hexahedra, etc.) simultaneously. Previously, the reader could only read
one or the other.

Note that in this situation, only the mesh is read, not the flow solutions
and boundary conditions.