vtkThreadedCallbackQueue: removing Start and Stop from public API
This API can be dangerous because the user could call `Wait()` on a
token an end up in a deadlock if `Stop()` was called anywhere.
`Stop` and `Start` are now private.

There also was a bug in `SetNumberOfThreads`: the thread id was not
instantiated correctly when resizing the threads when the queue was
already running.