vtkThreadedCallbackQueue: replacing Token by SharedFuture
Having a `std::shared_future` instance made that independent locks were
done when they could have been combined. The `vtkCallbackToken` was
actually already near to have all the functionalities of a
`std::shared_future`.

We also wanted to avoid `std::shared_future::wait_for(0)`, as this call
blocks the current thread. In our implementation, if the invoker is not
done yet, you can call an equivalent of `wait_for(0)` as `IsReady`
without actually blocking the thread.