Apply try/except to importing most python modules
In order to make it easier for package managers to build VTK with
everything and then split it into smaller binary packages, use
a try/except for importing all but the StandAlone and Rendering
modules.  This allows "import vtk" to work even if not all of the
built modules are installed.