Using annular requests for point neighbors
Reworked the algorithm by redefing the process of identifying
nearby points to the Voronoi tile generating point. Now disjoint,
annular requests are made until the Voronoi Flower error metric is
satisfied.

Additionally, some basic infrastructure was put in place to support
the generation of the Delaunay triangulation from the Voronoi
tessellation. This includes defining a hub-and-spoke data structure,
and adding data members to turn on and off the generation of the
Delaunay triangulation.