Added region support for Voronoi tiles
Region ids can not be provided for each input point (i.e., tile
generator). These ids are supplied through an input point data array.
A region id == 0 means that the point should be considered "outside"
so the tile is not sent to the output. Non-zero region ids serve as
a labeling to the resulting output tiles.