Add UseImplicitArrays option in vtk(P)HTGProbeFilter
vtkHyperTreeGridProbeFilter now exposes the UseImplictArray option.
When on, the filter will use indexed arrays internally to improve the memory consumption of the filter,
at the price of a higher computational cost.

Even if changes have also been made in the vtkPHyperTreeGridProbeFilter in this direction,
this option should not be used on this class until vtkHyperTreeGrid fully supports global IDs.