vtkModule: name targets after the library name of modules
This naming avoids conflicts when, e.g., a base name of a module is
repeated in different namespaces (ParaView plugin inner modules
sometimes do this). It also avoids the problem when an external
dependency uses the base name of a third party module as its target
(which is becoming more common).

Fixes: #17679