Update STL writer test for proper empty data writing
Previously, the STL writer would fail to write data that had no points.
This was because the `Points` member was set to a null `vtkPoints`
object. This behavior has changed so that `vtkSTLWriter` correctly
writes out an empty .stl file. Modify the associated test to check for
this better behavior.