Correct test data for TestVTKMNDHistogram
Previously, the test did not set any points. Through a couple of fragile
behaviors, no coordinates were set on the VTK-m DataSet and the number
of points was reported from the point fields. Because `vtkPointSet` no
lonfger allows for null points, the VTK-m DataSet got set an empty
coordinat system, which caused problems with the filter.