Remove testing of diagonal in empty generic data set
In a test of the behavior of an empty generic data set, it tries to
check the length of the diagonal. Before this check, it already verifies
that the bounds are unitialized. After that, it assumes the length is
2sqrt(3), which is a nonsensical result of a happenstance. The
happenstance has changed and the test started failing for no reason.
Just delete this nonsense test.